services/domain-analyze/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
asyncpg
alembic
httpx
celery
